refactor: change type to typescript
