+++ packages/graphql-generator/src/index.ts
@@ -2,4 +2,12 @@ export { generateModels } from './models';
 export { generateStatements } from './statements';
 export { generateTypes } from './types';
 
-export type { TypesTarget, ModelsTarget, StatementsTarget, Language, FileExtension, GenerateTypesOptions, GenerateModelsOptions, GenerateStatementsOptions, GeneratedOutput } from './type';
+export type {
+  TypesTarget,
+  ModelsTarget,
+  StatementsTarget,
+  GenerateTypesOptions,
+  GenerateModelsOptions,
+  GenerateStatementsOptions,
+  GeneratedOutput
+} from './typescript';
